Insert.py Add Insert operation for BST
Once the Search operation identifies the node the Insert operation is performed. A new key is always inserted at the leaf node. The Search key starts searching a key from root till the leaf node is hit. Once a leaf node is found, the node is added as a child of the leaf node. 

Closes https://github.com/fnplus/interview-techdev-guide/issues/165