kruskals_algorithm.cpp: Add Kruskals Algorithm
This adds Kruskals Algorithm, a minimum-spanning-tree algorithm, which finds an edge of the least possible weight that connects any two trees in the forest. It is a greedy algorithm in graph theory as it finds a minimum spanning tree for a connected weighted graph adding increasing cost arcs at each step.