Add Binary Search Algorithm
https://en.wikipedia.org/wiki/Binary_search_algorithm

This finds the position of a target value within a sorted array. It compares the target value to the middle element of the array. If it is not equal the half that the target cannot be in is 'removed' and the search continues on the remaining half, again taking the middle element to compare to the target value, repeating this till the target value is found. If the search ends with the remaining half being empty, the target is not in the array.